extracting gitHubConfig
